Полина сделала Card и Section
